Chore: Const, Util 폴더 정리
- Util 폴더 Utils 로 통합
- Const 폴더 생성
    - IntentKeys object class 생성
