+++ app/src/main/java/com/example/spnapplication/const/IntentKeys.kt
@@ -1,0 +1,6 @@
+package com.example.spnapplication.const
+
+object IntentKeys {
+
+    const val USER_INFO = "user_info"
+}
